requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
python-dotenv